Fix Arrow deserialization for `FlatArray` (#29)
Tests currently fail with the latest compatible versions of this
package's dependencies because deserializing a `FlatArray` from Arrow
ends up calling a `FlatArray` constructor method that doesn't exist:
`FlatArray(vec, size)`. What it should actually call is the inner
constructor, `FlatArray{T}(vec, size)`. To fix this, we can define a
method for `fromarrow` from ArrowTypes to simply call the inner
constructor. An analogous change is not needed for `Weights` because the
method that `fromarrow` calls by default does exist in that case.

Note that no new tests were added. This is because the current tests
actually cover this case.

Co-authored-by: Angelica Patino <angelica.patino@beacon.bio>